+++ examples/rp2040/src/bin/display/mod.rs
@@ -1,4 +1,0 @@
-mod rect;
-mod writer;
-
-pub use writer::*;